组织账号管理功能

+++ application/common/helper/StringHelper.php
@@ -37,7 +37,7 @@ class StringHelper
     public static function leftPadSpace($name, $now_level, $begin_level = 0)
     {
         $pad_str = '&nbsp;';
-        $pad_num = floor(pow(($now_level - $begin_level - 1), 1.9) * 2);
+        $pad_num = (int)floor(pow(($now_level - $begin_level - 1), 1.9) * 2);
         if ($pad_num <= 0) {
             return $name;
         }
